+++ apps/student/src/main/java/com/instructure/student/features/dashboard/edit/itemviewmodels/EditDashboardEnrollmentItemViewModel.kt
@@ -14,25 +14,15 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.instructure.student.features.dashboard.edit.itemViewModel
+package com.instructure.student.features.dashboard.edit.itemviewmodels
 
 import androidx.annotation.StringRes
-import androidx.databinding.BaseObservable
-import androidx.databinding.Bindable
 import com.instructure.pandautils.mvvm.ItemViewModel
 import com.instructure.student.R
 import com.instructure.student.features.dashboard.edit.EditDashboardItemViewType
 
-class EditDashboardHeaderViewModel(@get:StringRes val title: Int, @get:Bindable var hasItemSelected: Boolean, val selectAllHandler: () -> Unit, val deselectAllHandler: () -> Unit) : ItemViewModel, BaseObservable() {
-    override val layoutId: Int = R.layout.viewholder_edit_dashboard_header
+class EditDashboardEnrollmentItemViewModel(@get:StringRes val title: Int) : ItemViewModel {
+    override val layoutId: Int = R.layout.viewholder_edit_dashboard_enrollment
 
-    override val viewType: Int = EditDashboardItemViewType.HEADER.viewType
-
-    fun onActionClick() {
-        if (hasItemSelected) {
-            deselectAllHandler()
-        } else {
-            selectAllHandler()
-        }
-    }
+    override val viewType: Int = EditDashboardItemViewType.ENROLLMENT.viewType
 }
