+++ apps/student/src/main/java/com/instructure/student/features/dashboard/edit/itemviewmodels/EditDashboardGroupItemViewModel.kt
@@ -14,40 +14,36 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.instructure.student.features.dashboard.edit.itemViewModel
+package com.instructure.student.features.dashboard.edit.itemviewmodels
 
 import androidx.databinding.BaseObservable
 import androidx.databinding.Bindable
 import com.instructure.pandautils.mvvm.ItemViewModel
-import com.instructure.student.R
 import com.instructure.student.features.dashboard.edit.EditDashboardItemAction
+import com.instructure.student.R
 import com.instructure.student.features.dashboard.edit.EditDashboardItemViewType
 
-class EditDashboardCourseItemViewModel(val id: Long, val name: String?, @get:Bindable var isFavorite: Boolean, val favoriteable: Boolean, val openable: Boolean, val termTitle: String?, private val actionHandler: (EditDashboardItemAction) -> Unit) : ItemViewModel, BaseObservable() {
+class EditDashboardGroupItemViewModel(
+        val id: Long,
+        val name: String?,
+        @get:Bindable var isFavorite: Boolean,
+        val subtitle: String?,
+        val termTitle: String?,
+        private val actionHandler: (EditDashboardItemAction) -> Unit
+) : ItemViewModel, BaseObservable() {
+    override val layoutId: Int = R.layout.viewholder_edit_dashboard_group
 
-    override val layoutId: Int = R.layout.viewholder_edit_dashboard_course
-
-    override val viewType: Int = EditDashboardItemViewType.COURSE.viewType
+    override val viewType: Int = EditDashboardItemViewType.GROUP.viewType
 
     fun onClick() {
-        if (!openable) {
-            actionHandler(EditDashboardItemAction.ShowSnackBar(R.string.unauthorized))
-            return
-        }
-
-        actionHandler(EditDashboardItemAction.OpenCourse(id))
+        actionHandler(EditDashboardItemAction.OpenGroup(id))
     }
 
     fun onFavoriteClick() {
-        if (!favoriteable) {
-            actionHandler(EditDashboardItemAction.ShowSnackBar(R.string.inactive_courses_cant_be_added_to_dashboard))
-            return
-        }
-
         if (isFavorite) {
-            actionHandler(EditDashboardItemAction.UnfavoriteCourse(this))
+            actionHandler(EditDashboardItemAction.UnfavoriteGroup(this))
         } else {
-            actionHandler(EditDashboardItemAction.FavoriteCourse(this))
+            actionHandler(EditDashboardItemAction.FavoriteGroup(this))
         }
     }
 }
