+++ apps/student/src/main/res/layout/viewholder_edit_dashboard_description.xml
@@ -20,7 +20,7 @@
 
         <variable
             name="itemViewModel"
-            type="com.instructure.student.features.dashboard.edit.itemViewModel.EditDashboardDescriptionItemViewModel" />
+            type="com.instructure.student.features.dashboard.edit.itemviewmodels.EditDashboardDescriptionItemViewModel" />
     </data>
 
     <TextView
