+++ apps/student/src/main/res/layout/viewholder_edit_dashboard_enrollment.xml
@@ -21,7 +21,7 @@
 
         <variable
             name="itemViewModel"
-            type="com.instructure.student.features.dashboard.edit.itemViewModel.EditDashboardEnrollmentItemViewModel" />
+            type="com.instructure.student.features.dashboard.edit.itemviewmodels.EditDashboardEnrollmentItemViewModel" />
     </data>
 
     <RelativeLayout
